// src/app/core/enums/category.ts
export enum CATEGORY {
  WOMEN_FASHION = 'WOMEN_FASHION',
  MEN_FASHION = 'MEN_FASHION',
  ELECTRONICS = 'ELECTRONICS',
  HOME_AND_LIFESTYLE = 'HOME_AND_LIFESTYLE',
  MEDICINE = 'MEDICINE',
  SPORTS_AND_OUTDOOR = 'SPORTS_AND_OUTDOOR',
  BABY_AND_TOYS = 'BABY_AND_TOYS',
  GROCERIES_AND_PETS = 'GROCERIES_AND_PETS',
  HEALTH_AND_BEAUTY = 'HEALTH_AND_BEAUTY',
}